Move sys._set_exception() to _contextlib._setup_genmgr_ctx
Since this is only used for a contextlib-specific hack, having it
exposed in the sys module (even with an underscore) might not be a good
idea.
