--- Misc/NEWS.d/next/Core_and_Builtins/2023-12-19-17-27-08.gh-issue-111375.XeHOl_.rst
@@ -1,2 +1,0 @@
-Add sys._set_exception() function that can set/clear the current exception
-context. Patch by Carey Metcalfe.
