+++ PCbuild/pythoncore.vcxproj.filters
@@ -989,6 +989,9 @@
     <ClCompile Include="..\Modules\_collectionsmodule.c">
       <Filter>Modules</Filter>
     </ClCompile>
+    <ClCompile Include="..\Modules\_contextlibmodule.c">
+      <Filter>Modules</Filter>
+    </ClCompile>
     <ClCompile Include="..\Modules\_csv.c">
       <Filter>Modules</Filter>
     </ClCompile>
